Separate monitor from compiler code
