+++ pyp5js/compiler.py
@@ -1,11 +1,8 @@
 import shlex
 import shutil
 import subprocess
-import time
 from cprint import cprint
 from unipath import Path
-from watchdog.events import PatternMatchingEventHandler
-from watchdog.observers import Observer
 
 from pyp5js.fs import Pyp5jsLibFiles
 
@@ -62,40 +59,3 @@ class Pyp5jsCompiler:
 def compile_sketch_js(sketch_files):
     compiler = Pyp5jsCompiler(sketch_files)
     compiler.compile_sketch_js()
-
-
-def monitor_sketch(sketch_files):
-    event_handler = TranscryptSketchEventHandler(sketch_files=sketch_files)
-    observer = Observer()
-
-    observer.schedule(event_handler, sketch_files.sketch_dir)
-    observer.start()
-    try:
-        while True:
-            time.sleep(1)
-    except KeyboardInterrupt as e:
-        observer.stop()
-        raise e
-    observer.join()
-
-
-class TranscryptSketchEventHandler(PatternMatchingEventHandler):
-    patterns = ["*.py"]
-
-    def __init__(self, *args, **kwargs):
-        self.sketch_files = kwargs.pop('sketch_files')
-        self._last_event = None
-        super().__init__(*args, **kwargs)
-
-    def on_modified(self, event):
-        event_id = id(event)
-        if event_id == self._last_event:
-            return
-
-        cprint.info(f"New change in {event.src_path}")
-
-        compile_sketch_js(self.sketch_files)
-        self._last_event = event_id
-
-        index_file = self.sketch_files.index_html
-        cprint.ok(f"Your sketch is ready and available at {index_file}")
